Allow setting of target repo checkout revision
Adds environment variable so that the target revision of dependency
repos can be set to something other than HEAD if needed. Helps prevent
churn / dependency failures with those repos are under active
development.

Note that the rev must be named, such as a branch or label. If it
wouldn't work with `git rev-parse`, then it won't work here.