Fix missing zenoh-c symbols in CI build
Address failing CI build (missing zenoh-c symbols) by switching to
no-as-needed linking for zenoh-c and using a different library search
technique in cmake.

Adding -Wl,--no-as-needed is necessary when linking an application
against a static library where *another* library is dependent on that
static library's symbols. In this case, up-client-zenoh-cpp depends on
zenoh-c, but does not incorporate those symbols directly. At link time,
the linker will look at the symbols provided by zenoh-c and the symbols
needed for the application binary, determine none are needed, and
discard them. The -Wl,--no-as-needed flag tells the linker to embed
those symbols into the application binary anyway.

It was also necessary to explicitly pass a search directory and library
name to the linker instead of providing cmake a path to a static library
file. It is unclear why providing a library path was not working - the
documentation indicates it should. However, using standard -L<path> and
-l<library> flags did resolve the issue.

I tested with each of the above solutions independently and neither one
worked on its own. Both were required to fix the CI build.

This change also re-enables build for examples that were previously
disabled.